Quieted implicit-conversion warnings
-Wshorten-64-to-32, -Wimplicit-int-conversion

+++ deps/uv_link_t/src/uv_link_t.c
@@ -177,7 +177,7 @@ void uv_link_propagate_close(uv_link_t* link, uv_link_t* source,
 
 
 static void uv_link_recalculate_prefixes(uv_link_t* link) {
-  unsigned short prev;
+  unsigned prev;
 
   prev = link->parent->err_prefix >> kErrorPrefixShift;
   for (; link != NULL; link = link->child)
